BD float

+++ persistenceBS/src/main/resources/schema.sql
@@ -14,6 +14,6 @@ CREATE TABLE IF NOT EXISTS hospitals (
 	hospitalId SERIAL PRIMARY KEY,
 	hospitalName VARCHAR(64) NOT NULL,
 	hospitalAddress VARCHAR(64) NOT NULL UNIQUE,
-	latitude DOUBLE NOT NULL,
-	longitude DOUBLE NOT NULL
+	latitude FLOAT NOT NULL,
+	longitude FLOAT NOT NULL
 );
